+++ servers/db/Dockerfile
@@ -1,3 +1,2 @@
 FROM mysql
-EXPOSE 3306
 COPY schema.sql /docker-entrypoint-initdb.d/schema.sql
