+++ servers/gateway/main.go
@@ -13,6 +13,7 @@ import (
 	"time"
 
 	"github.com/go-redis/redis"
+	_ "github.com/go-sql-driver/mysql"
 )
 
 //main is the main entry point for the server
