+++ servers/summary/Dockerfile
@@ -1,0 +1,4 @@
+FROM alpine
+COPY summary /summary
+EXPOSE 80
+ENTRYPOINT ["/summary"]
