+++ servers/summary/build.sh
@@ -1,0 +1,5 @@
+GOOS=linux go build
+
+docker build -t lyons124/summary .
+
+go clean
