+++ servers/summary/deploy.sh
@@ -1,0 +1,5 @@
+sh build.sh
+
+docker push lyons124/summary:latest
+
+ssh ec2-user@api.lyonsupernova.me < update.sh
