+++ servers/summary/main.go
@@ -1,0 +1,31 @@
+package main
+
+import (
+	summary "assignments-fixed-Lyonsupernova/servers/summary/handlers"
+	"log"
+	"net/http"
+	"os"
+)
+
+//main is the main entry point for the server
+func main() {
+	/* TODO: add code to do the following
+	- Read the ADDR environment variable to get the address
+	  the server should listen on. If empty, default to ":80"
+	- Create a new mux for the web server.
+	- Tell the mux to call your handlers.SummaryHandler function
+	  when the "/v1/summary" URL path is requested.
+	- Start a web server listening on the address you read from
+	  the environment variable, using the mux you created as
+	  the root handler. Use log.Fatal() to report any errors
+	  that occur when trying to start the web server.
+	*/
+
+	addr := os.Getenv("ADDR")
+	if len(addr) == 0 {
+		addr = ":80"
+	}
+	mux := http.NewServeMux()
+	mux.HandleFunc("/v1/summary", summary.SummaryHandler)
+	log.Fatal(http.ListenAndServe(addr, mux))
+}
